CS452 Realtime Micro Kernel
This is a project for CS452 at the University of Waterloo. It's a
result of the many hours William Saunders and I spent in the trains
lab in 2011.

Included is a realtime micro kernel, and application code to run atop
it for controlling a model train set. A simple game where the trains
represent "repair robots" which you must outsmart to destroy the
spaceship (represented by the track) is included.